GVT-2203 Use publication revert code for deleting track numbers
