+++ ui/src/tool-panel/track-number/track-number-deletion.ts
@@ -1,0 +1,23 @@
+import { getRevertRequestDependencies } from 'publication/publication-api';
+import { publishNothing } from 'publication/publication-model';
+import { PreviewSelectType } from 'preview/preview-table';
+import { LayoutTrackNumber } from 'track-layout/track-layout-model';
+import { ChangesBeingReverted } from 'preview/preview-view';
+
+export const onRequestDeleteTrackNumber = (
+    trackNumber: LayoutTrackNumber,
+    setChangesBeingReverted: (changes: ChangesBeingReverted) => void,
+) => {
+    getRevertRequestDependencies({ ...publishNothing, trackNumbers: [trackNumber.id] }).then(
+        (changeIncludingDependencies) =>
+            changeIncludingDependencies &&
+            setChangesBeingReverted({
+                requestedRevertChange: {
+                    type: PreviewSelectType.trackNumber,
+                    name: trackNumber.number,
+                    id: trackNumber.id,
+                },
+                changeIncludingDependencies,
+            }),
+    );
+};
